refactor(folder-structure): move theme related files to `theme` folder
