+++ src/theme/components/EditColorsModal.js
@@ -1,14 +1,11 @@
 import * as React from 'react'
-import { useTheme } from '@material-ui/core'
-import { ColorPalette } from 'material-ui-color'
-import { FaCircle, FaLandmark, FaMonument } from 'react-icons/fa'
 import Box from '@material-ui/core/Box'
 import Dialog from '@material-ui/core/Dialog'
 import Divider from '@material-ui/core/Divider'
 import DialogContent from '@material-ui/core/DialogContent'
 import DialogTitle from '@material-ui/core/DialogTitle'
-import Typography from '@material-ui/core/Typography'
 
+import SelectColorPane from './SelectColorPane'
 import palette from 'theme/palette'
 import { emit } from 'theme/mediator'
 
@@ -19,7 +16,7 @@ export default function ColorsModal({ onClose, open }) {
 
       <DialogContent dividers>
         <Box sx={{ display: 'flex', gap: 2 }}>
-          <ColorSelectPane
+          <SelectColorPane
             color="primary"
             title="Primary color"
             onSelect={color => emit('primaryColorSelected', palette[color])}
@@ -27,7 +24,7 @@ export default function ColorsModal({ onClose, open }) {
 
           <Divider flexItem orientation="vertical" />
 
-          <ColorSelectPane
+          <SelectColorPane
             color="secondary"
             title="Secondary color"
             onSelect={color => emit('secondaryColorSelected', palette[color])}
@@ -35,32 +32,5 @@ export default function ColorsModal({ onClose, open }) {
         </Box>
       </DialogContent>
     </Dialog>
-  )
-}
-
-function ColorSelectPane({ title, color, onSelect }) {
-  const theme = useTheme()
-
-  const iconStyle = {
-    color: theme.palette[color].main,
-    fontSize: 60,
-  }
-
-  return (
-    <Box sx={{ maxWidth: 200 }}>
-      <Typography align="center" fontWeight="bold" sx={{ mb: 1 }}>
-        {title}
-      </Typography>
-
-      <Box sx={{ mb: 3 }}>
-        <ColorPalette palette={palette} onSelect={onSelect} />
-      </Box>
-
-      <Box sx={{ display: 'flex', justifyContent: 'space-evenly' }}>
-        <FaLandmark style={iconStyle} />
-        <FaMonument style={iconStyle} />
-        <FaCircle style={iconStyle} />
-      </Box>
-    </Box>
   )
 }
