+++ src/theme/components/EditColorsToggler.js
@@ -2,10 +2,10 @@ import * as React from 'react'
 import IconButton from '@material-ui/core/IconButton'
 import InvertColorsIcon from '@material-ui/icons/InvertColors'
 
-import ColorsModal from './ColorsModal'
+import ColorsModal from '../theme/components/ColorsModal'
 import { Tooltip } from '@material-ui/core'
 
-export default function ColorsModalToggler() {
+export default function EditColorsToggler() {
   const [isOpen, setIsOpen] = React.useState(false)
   const openModal = () => setIsOpen(true)
   const closeModal = () => setIsOpen(false)
